Bring Client/Locations into the engine, use G5 Foundation Client.
This is a pretty enormous commit. I might have been able to split this
into two commits, but it would have been more trouble than it was worth
The microformats2 parsing bits were removed in favor of the
g5_foundation_client gem. Client and Location tables and ActiveRecord
models were created here in favor of having to add them to each
individual app. The need for most of the custom configuration stuff was
obviated by that decision, so that's ripped out. The only one still in
place is setting a client_uid, but that's never going to be set in
committed code anyway, so it uses an environment variable now.