Add Location#client method.
I'm not using a traditional belongs_to because I want to avoid using the
database IDs entirely. The job that reads from G5 Hub should be able to
destroy and create the client records without worrying about orphaning
locations, as long as the UID remains consistent (which is basically its
one job).